[prog] work on alternative filter dropdowns
